[api] removed unnecessary api calls
