+++ api/dapp_test.go
@@ -1,17 +1,42 @@
 package api
 
 import (
+	"crypto/rand"
+	"encoding/hex"
+	"fmt"
 	"testing"
 	"time"
 
-	"crypto/rand"
-	"encoding/hex"
 	pb "github.com/Bit-Nation/panthalassa/api/pb"
-	dapp "github.com/Bit-Nation/panthalassa/dapp"
+	keyManager "github.com/Bit-Nation/panthalassa/keyManager"
+	keyStore "github.com/Bit-Nation/panthalassa/keyStore"
+	mnemonic "github.com/Bit-Nation/panthalassa/mnemonic"
 	proto "github.com/golang/protobuf/proto"
 	require "github.com/stretchr/testify/require"
 	ed25519 "golang.org/x/crypto/ed25519"
 )
+
+func keyManagerFactory() *keyManager.KeyManager {
+
+	mne, err := mnemonic.FromString("panda eyebrow bullet gorilla call smoke muffin taste mesh discover soft ostrich alcohol speed nation flash devote level hobby quick inner drive ghost inside")
+	if err != nil {
+		panic(err)
+	}
+
+	ks, err := keyStore.NewFromMnemonic(mne)
+	if err != nil {
+		panic(err)
+	}
+
+	return keyManager.CreateFromKeyStore(ks)
+
+}
+
+func requireNil(value interface{}) {
+	if value != nil {
+		panic(fmt.Sprintf("Expected value: %s to be nil", value))
+	}
+}
 
 func TestAPI_ShowModal(t *testing.T) {
 
@@ -108,57 +133,4 @@ func TestAPI_SendEthereumTransaction(t *testing.T) {
 	resp, err := api.SendEthereumTransaction("100", "0x1f75bb626ad018f3354259b10ab2e74bc0e0f267", "0xf3")
 	require.Nil(t, err)
 	require.Equal(t, `{"chainId":4,"data":"0xf3","from":"my_address","gasLimit":"100000000000","gasPrice":"1000000000","hash":"tx-hash","nonce":3,"r":"r_of_tx","s":"s_of_tx","to":"0x1f75bb626ad018f3354259b10ab2e74bc0e0f267","v":"v_of_tx","value":"100"}`, resp)
-}
-
-func TestAPI_SaveDApp(t *testing.T) {
-
-	c := make(chan string)
-
-	api := New(&testUpStream{
-		sendFn: func(data string) {
-			c <- data
-		},
-	}, keyManagerFactory())
-
-	go func() {
-
-		select {
-		case data := <-c:
-
-			req := pb.Request{}
-			requireNil(proto.Unmarshal([]byte(data), &req))
-
-			app := req.SaveDApp
-
-			if "My DApp" != app.AppName {
-				panic("Expected dapp name to be: My DApp")
-			}
-
-			if "var i = 3" != app.Code {
-				panic("Expected dapp name to be: var i = 3")
-			}
-
-			if "7075626b6579" != app.SigningPublicKey {
-				panic("Expected dapp name to be: My DApp")
-			}
-
-			if "7369676e6174757265" != app.Signature {
-				panic("Expected dapp name to be: My DApp")
-			}
-
-			err := api.Respond(req.RequestID, &pb.Response{}, nil, time.Second*5)
-			if err != nil {
-				panic(err)
-			}
-		}
-
-	}()
-
-	err := api.SaveDApp(dapp.JsonRepresentation{
-		Name:               "My DApp",
-		Code:               "var i = 3",
-		SignaturePublicKey: []byte("pubkey"),
-		Signature:          []byte("signature"),
-	})
-	require.Nil(t, err)
 }
