+++ api/pb/request.proto
@@ -6,39 +6,6 @@ message Request {
 
     string requestID = 1;
 
-    message DRKeyStoreGet {
-        bytes drKey = 1;
-        uint64 messageNumber = 2;
-    }
-
-    message DRKeyStorePut {
-        bytes key = 1;
-        uint64 messageNumber = 2;
-        bytes messageKey = 3;
-    }
-
-    message DRKeyStoreDeleteMK {
-        bytes key = 1;
-        uint64 msgNum = 2;
-    }
-
-    message DRKeyStoreDeleteKeys {
-        bytes key = 1;
-    }
-
-    message DRKeyStoreCount {
-        bytes key = 1;
-    }
-
-    message DRKeyStoreAll {}
-
-    DRKeyStoreGet dRKeyStoreGet = 2;
-    DRKeyStorePut dRKeyStorePut = 3;
-    DRKeyStoreDeleteMK dRKeyStoreDeleteMK = 4;
-    DRKeyStoreDeleteKeys dRKeyStoreDeleteKeys = 5;
-    DRKeyStoreCount dRKeyStoreCount = 6;
-    DRKeyStoreAll dRKeyStoreAll = 7;
-
     message ShowModal {
         bytes dAppPublicKey = 1;
         string title = 2;
@@ -54,14 +21,5 @@ message Request {
     }
 
     SendEthereumTransaction sendEthereumTransaction = 9;
-
-    message SaveDApp {
-        string appName = 1;
-        string code = 2;
-        string signature = 3;
-        string signingPublicKey = 4;
-    }
-
-    SaveDApp saveDApp = 10;
 
 }
