+++ api/pb/response.proto
@@ -4,29 +4,6 @@ package api_proto;
 
 message Response {
 
-    message DRKeyStoreGet {
-        bytes messageKey = 1;
-    }
-
-    message DRKeyStoreCount {
-        uint64 count = 1;
-    }
-
-    message DRKeyStoreAll {
-
-        message Key {
-            bytes key = 1;
-            map<uint64, bytes> messageKeys = 2;
-        }
-
-        repeated Key all = 1;
-
-    }
-
-    DRKeyStoreGet dRKeyStoreGet = 3;
-    DRKeyStoreCount dRKeyStoreCount = 4;
-    DRKeyStoreAll dRKeyStoreAll = 5;
-
     message SendEthereumTransaction {
         uint32 nonce = 1;
         // must be base 10!
